+++ src/view/card-film/card-film.js
@@ -1,13 +1,15 @@
-import {createElement} from '../utils.js';
-import {createFilmsTemplate} from './template.js';
+import {createElement} from '../../utils.js';
+import {createCardFilmTemplate} from './card-film-tpl';
 
-export default class Films {
-  constructor() {
+export default class CardFilm {
+  constructor(film) {
+    this._film = film;
     this._element = null;
   }
 
   getTemplate() {
-    return createFilmsTemplate();
+
+    return createCardFilmTemplate(this._film);
   }
 
   getElement() {
