+++ src/view/no-films/no-films-tpl.js
@@ -1,0 +1,9 @@
+export const createNoFilmTemplate = () => (
+  `<section class="films">
+    <section class="films-list">
+      <h2 class="films-list__title">
+        There are no movies in our database
+      </h2>
+    </section>
+  </section>`
+);
