+++ src/view/popup/popup-tpl.js
@@ -1,129 +1,8 @@
-import {isClassName, ucFirstName, isClassNamePopup} from '../utils.js';
-import {LENGTH_STRING, RANK} from '../mock/data.js';
+import {isClassNamePopup} from '../../utils.js';
 import dayjs from 'dayjs';
 import relativeTime from 'dayjs/plugin/relativeTime.js';
 
 dayjs.extend(relativeTime);
-
-export const createFilmsTemplate = () => (
-  `<section class="films">
-    <section class="films-list">
-      <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-
-      <div class="films-list__container">
-
-      </div>
-    </section>
-    <section class="films-list films-list--extra">
-      <h2 class="films-list__title">Top rated</h2>
-      <div class="films-list__container">
-
-      </div>
-    </section>
-    <section class="films-list films-list--extra films-list--most-commented">
-      <h2 class="films-list__title">Most commented</h2>
-      <div class="films-list__container">
-
-      </div>
-    </section>
-  </section>`
-);
-
-const isNameRank = (count) => {
-  let nameRank = 'novice';
-  if (count >= RANK.FAN.MIN && count <= RANK.FAN.MAX) {
-    nameRank = 'fan';
-  } else if (count >= RANK.MOVIE_BUFF) {
-    nameRank = 'movie buff';
-  }
-  return ucFirstName(nameRank);
-};
-
-const isGenerateProfile = (count) => (count === 0) ? ' ' : `<p class="profile__rating">${isNameRank(count)}</p><img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">`;
-
-export const createRankUserTemplate = (filter) => {
-  const count = filter[0].count;
-  return `<section class="header__profile profile">
-      ${isGenerateProfile(count)}
-    </section>`;
-};
-
-
-const createMenuTemplate = (filter) => {
-  const {name, count} = filter;
-  const ucFirstNameFilter = ucFirstName(name);
-  return `<a href="#${name}" class="main-navigation__item">${ucFirstNameFilter}<span class="main-navigation__item-count">${count}</span></a>`;
-};
-
-export const createFilterTemplate = (filterItems) => {
-  const filterItemsTemplate = filterItems
-    .map((filter, index) => createMenuTemplate(filter, index === 0))
-    .join('');
-
-  return `<nav class="main-navigation">
-      <div class="main-navigation__items">
-        <a href="#all" class="main-navigation__item main-navigation__item--active">All movies</a>
-          ${filterItemsTemplate}
-      </div>
-      <a href="#stats" class="main-navigation__additional">Stats</a>
-    </nav>`;
-};
-
-export const createSortTemplate = () => (
-  `<ul class="sort">
-    <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
-    <li><a href="#" class="sort__button">Sort by date</a></li>
-    <li><a href="#" class="sort__button">Sort by rating</a></li>
-  </ul>`
-);
-
-export const createNoFilmTemplate = () => (
-  `<section class="films">
-    <section class="films-list">
-      <h2 class="films-list__title">
-        There are no movies in our database
-      </h2>
-    </section>
-  </section>`
-);
-
-
-export const createCardFilmTemplate = (card) => {
-  const {film} = card;
-  const {filmInfo, comments, userDetails} = film;
-  const {title, totalRating, releaseFilm, runtime, genre, poster, description} = filmInfo;
-  const {date} = releaseFilm;
-  const {watchList, alreadyWatched, favorite} = userDetails;
-
-  const dateYear = dayjs(date).format('YYYY');
-  const descriptionShort = (text) => text.length >= LENGTH_STRING ? `${text.substr(0, LENGTH_STRING)} <span class='more'>...</span>` : text;
-
-  return `<article class="film-card">
-    <h3 class="film-card__title">${title}</h3>
-    <p class="film-card__rating">${totalRating}</p>
-    <p class="film-card__info">
-      <span class="film-card__year">${dateYear}</span>
-      <span class="film-card__duration">${runtime}</span>
-      <span class="film-card__genre">${genre}</span>
-    </p>
-    <img src="${poster}" alt="" class="film-card__poster">
-    <p class="film-card__description">${descriptionShort(description)}</p>
-    <a class="film-card__comments">${comments.length} comments</a>
-    <div class="film-card__controls">
-      <button class="film-card__controls-item film-card__controls-item--add-to-watchlist ${isClassName(alreadyWatched)}" type="button">Add to watchlist</button>
-      <button class="film-card__controls-item film-card__controls-item--mark-as-watched ${isClassName(watchList)}" type="button">Mark as watched</button>
-      <button class="film-card__controls-item film-card__controls-item--favorite ${isClassName(favorite)}" type="button">Mark as favorite</button>
-    </div>
-  </article>`;
-};
-
-export const createShowMoreButtonTemplate = () => '<button class="films-list__show-more">Show more</button>';
-
-export const createStatisticTemplate = (length) => (
-  `<section class="footer__statistics">
-    <p>${length} movies inside</p>
-   </section>`
-);
 
 const isGenre = (array) => array.length > 1 ? 'Genres' : 'Genre';
