+++ src/view/statistic/statistic-tpl.js
@@ -1,0 +1,5 @@
+export const createStatisticTemplate = (length) => (
+  `<section class="footer__statistics">
+    <p>${length} movies inside</p>
+   </section>`
+);
